requirements.txt
ollama==0.2.0
qq_botpy==1.2.1
func_timeout==4.3.5
